+++ packages/typescript-plugin/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @tsrx/typescript-plugin
 
+## 0.4.7
+
+No changes in this release.
+
 ## 0.4.6
 
 ### Patch Changes